Move ShowMeQuestionComponent to office component namespace
