+++ src/pages/office/__tests__/office.spec.ts
@@ -35,7 +35,7 @@ import { MockComponent } from 'ng-mocks';
 import { RouteNumberComponent } from '../components/route-number/route-number';
 import { CandidateDescriptionComponent } from '../components/candidate-description/candidate-description';
 import { DebriefWitnessedComponent } from '../components/debrief-witnessed/debrief-witnessed';
-import { ShowMeQuestionComponent } from '../show-me-question/show-me-question';
+import { ShowMeQuestionComponent } from '../components/show-me-question/show-me-question';
 import { WeatherConditionsComponent } from '../components/weather-conditions/weather-conditions';
 import { D255Component } from '../components/d255/d255';
